refactor(db): fold db_* siblings into src/db/ namespace
Closes #75. Collapse the six prefix-named sibling modules
(db, db_job, db_models, db_pending, db_query, db_usage) into a single
src/db/ package so the DB layer is an actual boundary instead of a
filename prefix.

- src/db_job{.rs,/} -> src/db/job{.rs,/}
- src/db_models.rs   -> src/db/models.rs
- src/db_pending{.rs,/} -> src/db/pending{.rs,/}
- src/db_query{.rs,/}   -> src/db/query{.rs,/}
- src/db_usage.rs    -> src/db/usage.rs

Per U-24, no `pub use crate::db_*` aliases are kept; lib.rs no longer
declares the sibling modules and every caller now uses
`crate::db::<sub>::...`. The flat public surface re-exports inside
src/db.rs are updated to match (`pub use job::*`, etc.).

This makes `rg "crate::db::"` answer the question "does X reach into
the DB layer?" and lets future layer enforcement (cargo deny rules,
visibility audits) work against a single namespace.
