+++ src/db.rs
@@ -1,17 +1,21 @@
 pub mod core;
 pub mod crypto;
+pub mod job;
+pub mod models;
 pub mod observation;
+pub mod pending;
+pub mod query;
 pub mod summarize;
+pub mod usage;
 #[cfg(test)]
 pub mod test_support;
 
-pub use crate::db_job::*;
-pub use crate::db_models::*;
-pub use crate::db_pending::*;
-pub use crate::db_query::*;
-pub use crate::db_usage::*;
-
 pub use core::*;
 pub use crypto::*;
+pub use job::*;
+pub use models::*;
 pub use observation::*;
+pub use pending::*;
+pub use query::*;
 pub use summarize::*;
+pub use usage::*;
