+++ src/entity/graph/seed.rs
@@ -17,5 +17,5 @@ pub(super) fn load_seed_entity_ids(conn: &Connection, seed_memory_ids: &[i64]) -
         params_vec.iter().map(|value| value.as_ref()).collect();
     let mut stmt = conn.prepare(&sql)?;
     let rows = stmt.query_map(refs.as_slice(), |row| row.get::<_, i64>(0))?;
-    crate::db_query::collect_rows(rows)
+    crate::db::query::collect_rows(rows)
 }
