+++ src/lib.rs
@@ -14,11 +14,6 @@ pub mod claude_memory;
 pub mod cli;
 pub mod context;
 pub mod db;
-pub mod db_job;
-pub mod db_models;
-pub mod db_pending;
-pub mod db_query;
-pub mod db_usage;
 pub mod dedup;
 pub mod doctor;
 pub mod dream;
