+++ src/memory_format.rs
@@ -5,7 +5,7 @@ mod parse;
 mod tests;
 mod types;
 
-pub use crate::db_models::OBSERVATION_TYPES;
+pub use crate::db::models::OBSERVATION_TYPES;
 pub use escape::{xml_escape_attr, xml_escape_text};
 pub use extract::extract_field;
 pub use parse::parse_observations;
