+++ src/timeline/detail.rs
@@ -4,7 +4,7 @@ use anyhow::Result;
 use rusqlite::Connection;
 
 use crate::db::to_sql_refs;
-use crate::db_query::push_project_filter;
+use crate::db::query::push_project_filter;
 
 use super::types::{MonthRow, RecentObservation};
